Make each class and module has it's own file
