+++ main.rb
@@ -1,0 +1,10 @@
+require_relative 'my_list'
+
+  # main.rb
+  list = MyList.new(1, 2, 3, 4)
+  puts(list.all? { |e| e < 5 })
+  puts(list.all? { |e| e > 5 })
+  puts(list.any? { |e| e == 2 })
+  puts(list.any? { |e| e == 5 })
+  puts(list.filter(&:even?))
+
