+++ my_enumerable.rb
@@ -22,22 +22,3 @@ module MyEnumerable
     result
   end
 end
-
-# my_list.rb
-class MyList
-  include MyEnumerable
-  def initialize(*list)
-    @list = list
-  end
-
-  def each(&block)
-    @list.each(&block)
-  end
-end
-# main.rb
-list = MyList.new(1, 2, 3, 4)
-puts(list.all? { |e| e < 5 })
-puts(list.all? { |e| e > 5 })
-puts(list.any? { |e| e == 2 })
-puts(list.any? { |e| e == 5 })
-puts(list.filter(&:even?))
